Split View in ViewBase & View so that one can leverage ViewBase model & definition setup while not use the default templating mechanism from View (help a lot allowing people to plug an existing Widget as a View).
