+++ tests/customApp/DtlView.js
@@ -1,5 +1,5 @@
-define(["dojo/_base/lang", "dojo/_base/declare", "dojox/app/View", "dojox/dtl/_Templated"],
-	function(lang, declare, View, _Templated){
+define(["dojo/_base/declare", "dojox/app/View", "dojox/dtl/_Templated"],
+	function(declare, View, _Templated){
 		return declare([_Templated, View], {
 			_dijitTemplateCompat: true
 		});
