+++ tests/customApp/home.html
@@ -1,0 +1,10 @@
+<div class="view mblView">
+    <ul data-dojo-type="dojox/mobile/RoundRectList">
+        <li data-dojo-type="dojox/mobile/ListItem" clickable="true" transitionOptions="{title:'dtl',target:'dtl',url:'#dtl'}">
+           DTL Custom View
+        </li>
+        <li data-dojo-type="dojox/mobile/ListItem" clickable="true" transitionOptions="{title:'widget',target:'widget',url:'#widget'}">
+			Widget Custom View
+        </li>
+    </ul>
+</div>
