+++ tests/customApp/widget.js
@@ -1,0 +1,5 @@
+define([], function(){
+	return {
+		label: "test label"
+	}
+})
